requirements.txt
openai
python-docx
PyPDF2
faiss-cpu
numpy
pinecone-client
python-dotenv